requirements.txt
requests
python-dotenv
pandas
fastapi
uvicorn
pydantic
